requirements.txt
python-dotenv
pyyaml
pydantic
numpy
scikit-learn
rich
PyPDF2
langdetect
